fix(settings): reset loading state before router.refresh() in deletion actions
